2021.02.05 | 3015 | 오아시스 재결합

+++ Solution/03015_오아시스 재결합/BOJ_3015_오아시스 재결합.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int N;
 ll Ans;
-stack<int> Heights, Weights;
+stack<ll> Heights, Weights;
 
 int main() {
